defining list.js
