+++ server/gql/models/students/query.js
@@ -1,39 +1,40 @@
 import db from '@database/models';
 import { sequelizedWhere } from '@server/database/dbUtils';
-import { getNode } from '@server/gql/node';
+// import { getNode } from '@server/gql/node';
 import { totalConnectionFields } from '@server/utils';
-import { getQueryFields, TYPE_ATTRIBUTES } from '@server/utils/gqlFieldUtils';
-import { GraphQLID, GraphQLInt, GraphQLNonNull, GraphQLObjectType, GraphQLString } from 'graphql';
+// import { getQueryFields, TYPE_ATTRIBUTES } from '@server/utils/gqlFieldUtils';
+// import { GraphQLID, GraphQLInt, GraphQLNonNull, GraphQLObjectType, GraphQLString } from 'graphql';
 import { createConnection } from 'graphql-sequelize';
-import { limitAndOffset } from '../../fields/limitAndOffset';
-import { subjectQueries } from '../subjects';
-import { timestamps } from '../../fields/timestamps';
+// import { limitAndOffset } from '../../fields/limitAndOffset';
+// import { subjectQueries } from '../subjects';
+// import { timestamps } from '../../fields/timestamps';
+import { Student } from '.';
 
-const { nodeInterface } = getNode();
+// const { nodeInterface } = getNode();
 
-export const studentsFields = {
-  id: { type: new GraphQLNonNull(GraphQLID) },
-  name: { type: GraphQLString }
-};
+// export const studentsFields = {
+//   id: { type: new GraphQLNonNull(GraphQLID) },
+//   name: { type: GraphQLString }
+// };
 
-export const student = new GraphQLObjectType({
-  name: 'Student',
-  interfaces: [nodeInterface],
-  fields: () => ({
-    ...getQueryFields(studentsFields, TYPE_ATTRIBUTES.isNonNull),
-    ...timestamps,
-    subjects: {
-      ...subjectQueries.list,
-      resolve: (source, args, context, info) =>
-        subjectQueries.list.resolve(source, args, { ...context, student: source.dataValues }, info)
-    }
-  })
-});
+// export const student = new GraphQLObjectType({
+//   name: 'Student',
+//   interfaces: [nodeInterface],
+//   fields: () => ({
+//     ...getQueryFields(studentsFields, TYPE_ATTRIBUTES.isNonNull),
+//     ...timestamps,
+//     subjects: {
+//       ...subjectQueries.list,
+//       resolve: (source, args, context, info) =>
+//         subjectQueries.list.resolve(source, args, { ...context, student: source.dataValues }, info)
+//     }
+//   })
+// });
 
 export const StudentConnection = createConnection({
   name: 'students',
   target: db.students,
-  nodeType: student,
+  nodeType: Student,
   before: (findOptions, args, context) => {
     findOptions.include = findOptions.include || [];
 
@@ -52,23 +53,23 @@ export const StudentConnection = createConnection({
   ...totalConnectionFields
 });
 
-export const studentQuery = {
-  args: {
-    id: {
-      type: new GraphQLNonNull(GraphQLInt)
-    }
-  },
-  query: {
-    type: student
-  },
-  list: {
-    ...StudentConnection,
-    resolve: StudentConnection.resolve,
-    type: StudentConnection.connectionType,
-    args: {
-      ...StudentConnection.connectionArgs,
-      ...limitAndOffset
-    }
-  },
-  model: db.students
-};
+// export const studentQuery = {
+//   args: {
+//     id: {
+//       type: new GraphQLNonNull(GraphQLInt)
+//     }
+//   },
+//   query: {
+//     type: Student
+//   },
+//   list: {
+//     ...StudentConnection,
+//     resolve: StudentConnection.resolve,
+//     type: StudentConnection.connectionType,
+//     args: {
+//       ...StudentConnection.connectionArgs,
+//       ...limitAndOffset
+//     }
+//   },
+//   model: db.students
+// };
